+++ src/create-safe-route-handler.test.ts
@@ -42,12 +42,25 @@ describe('Extends Request type gracefully - convenience purpose for extending Ne
       }
     }
 
-    const GET = createSafeRouteHandler<NRequest>({}, async (_ctx, req) => {
-      expectTypeOf(req).toEqualTypeOf<NRequest>()
-      expect(req.cookies()).toBe('cookies')
+    const GET = createSafeRouteHandler(
+      {
+        id: 'extends-request-type',
+        authorize: async ({ req }) => {
+          expectTypeOf(req).toEqualTypeOf<Request>()
+          return { user: 'John Doe' }
+        },
+      },
+      async (ctx, req: NRequest) => {
+        expectTypeOf(req).toEqualTypeOf<NRequest>()
+        expect(req.cookies()).toBe('cookies')
 
-      return Response.json({ message: 'Hello, world!' }, { status: 200 })
-    })
+        expectTypeOf(ctx.auth).toEqualTypeOf<{
+          user: string
+        }>()
+        expect(ctx.auth.user).toBe('John Doe')
+        return Response.json({ message: 'Hello, world!' }, { status: 200 })
+      }
+    )
     const request = new NRequest('http://localhost:3000/')
     const response = await GET(request, { params: undefined })