+++ src/create-safe-route-handler.ts
@@ -76,15 +76,14 @@ const readRequestBodyAsJson = async <TReq extends Request>(
  * ```
  */
 export function createSafeRouteHandler<
-  TReq extends Request = Request,
   AC extends AuthContext | undefined = undefined,
   TRouteDynamicSegments extends TSegmentsDict | undefined = undefined,
   TSearchParams extends TSearchParamsDict | undefined = undefined,
   TBody extends TBodySchema | undefined = undefined,
   TFormData extends TFormDataDict | undefined = undefined,
+  TReq extends Request = Request,
 >(
   options: CreateSafeRouteHandlerOptions<
-    TReq,
     AC,
     TRouteDynamicSegments,
     TSearchParams,
@@ -92,12 +91,12 @@ export function createSafeRouteHandler<
     TFormData
   >,
   handlerFn: SafeRouteHandler<
-    TReq,
     AC,
     TRouteDynamicSegments,
     TSearchParams,
     TBody,
-    TFormData
+    TFormData,
+    TReq
   >
 ): CreateSafeRouteHandlerReturnType<TReq> {
   // NOTE: `body` and `formData` options are mutually exclusive 🎭
@@ -168,8 +167,9 @@ export function createSafeRouteHandler<
     const url = new URL(req.url)
 
     // Do not mutate / consume the original request
+    // Due to `NextRequest` limitations as the req is cloned it's always a Request
     const clonedReq_forAuth = req.clone() as TReq
-    const authOrResponse = await authorize({ req: clonedReq_forAuth, url })
+    const authOrResponse = await authorize({ url, req: clonedReq_forAuth })
     if (authOrResponse instanceof Response) {
       log.error(`🛑 Request not authorized for route handler '${id}'`)
       return authOrResponse